Set maximum heap size for JM. Rename POP<primitive> to POPMutable<primitive> (not POPString yet since it would impact more than the JobManager).
